Update Playwright test setup and refactor tests
This commit updates the Playwright test setup and refactors some existing tests to adapt to the changes.

For the test setup, a global setup file `globalSetup.ts` was added to maintain system information. The Playwright configuration is also updated to use `dotenv` for managing environment variables, and the test running scripts were modified.

In terms of tests, `example.spec.ts` was removed as it was no longer appropriate for the current project status. New tests were added in `qauto.spec.ts` and `auth.setup.ts` to cover more features of the application.

Lastly, `package-lock.json` and `package.json` were updated to reflect the addition of new dependencies. `.gitignore` was also updated to ignore `.env` and `.auth/` folders, which are used for storing sensitive data.