Refactor LoginPage and update auth setup
Refactored "LoginPage.ts" to extend from a newly created "BasePage.ts" class. Introduced a BasePage class for managing page-level actions such as navigation and added a signIn method in "LoginPage.ts" to encapsulate the authentication process. This simplifies and abstracts the authentication process in "auth.setup.ts". It improves readability and reusability of code across the application.